Push backing code for ReactSharedInternals into shared/. Next step is to push the __SECRET_INTERNALS symbol down to there as well, and remove the workaround for the circular dependency in the setupHostConfig and other places.
