+++ packages/react-reconciler/src/ReactFiberLane.js
@@ -8,30 +8,7 @@
  */
 
 import type {FiberRoot, ReactPriorityLevel} from './ReactInternalTypes';
-
-export opaque type LanePriority =
-  | 0
-  | 1
-  | 2
-  | 3
-  | 4
-  | 5
-  | 6
-  | 7
-  | 8
-  | 9
-  | 10
-  | 11
-  | 12
-  | 13
-  | 14
-  | 15
-  | 16
-  | 17;
-export opaque type Lanes = number;
-export opaque type Lane = number;
-export opaque type LaneMap<T> = Array<T>;
-
+import type {Lane, LanePriority, Lanes, LaneMap} from 'shared/ReactInternalTypes';
 import invariant from 'shared/invariant';
 
 import {
