+++ packages/react/src/ReactCurrentOwner.js
@@ -7,20 +7,3 @@
  * @flow
  */
 
-import type {Fiber} from 'react-reconciler/src/ReactInternalTypes';
-
-/**
- * Keeps track of the current owner.
- *
- * The current owner is the component who should own any components that are
- * currently being constructed.
- */
-const ReactCurrentOwner = {
-  /**
-   * @internal
-   * @type {ReactComponent}
-   */
-  current: (null: null | Fiber),
-};
-
-export default ReactCurrentOwner;
