+++ packages/shared/ReactComponentStackFrame.js
@@ -10,7 +10,7 @@
 import type {Source} from 'shared/ReactElementType';
 import type {LazyComponent} from 'react/src/ReactLazy';
 
-import {enableComponentStackLocations} from 'shared/ReactFeatureFlags';
+import {enableComponentStackLocations} from './ReactFeatureFlags';
 
 import {
   REACT_SUSPENSE_TYPE,
@@ -21,11 +21,9 @@ import {
   REACT_LAZY_TYPE,
 } from 'shared/ReactSymbols';
 
-import {disableLogs, reenableLogs} from 'shared/ConsolePatchingDev';
+import {disableLogs, reenableLogs} from './ConsolePatchingDev';
 
-import ReactSharedInternals from 'shared/ReactSharedInternals';
-
-const {ReactCurrentDispatcher} = ReactSharedInternals;
+import ReactCurrentDispatcher from './ReactSharedInternals';
 
 let prefix;
 export function describeBuiltInComponentFrame(
