+++ packages/shared/src/IsSomeRendererActing.js
