+++ packages/shared/src/ReactCurrentBatchConfig.js
