+++ packages/shared/src/ReactCurrentOwner.js
@@ -1,0 +1,16 @@
+import type {Fiber} from 'react-reconciler/src/ReactInternalTypes';
+
+/**
+ * Keeps track of the current owner.
+ *
+ * The current owner is the component who should own any components that are
+ * currently being constructed.
+ */
+const ReactCurrentOwner = {
+    /**
+     * @internal
+     * @type {ReactComponent}
+     */
+    current: (null: null | Fiber),
+};
+export default ReactCurrentOwner;
