+++ packages/shared/src/ReactInternalTypes.js
@@ -1,0 +1,50 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+import type {
+  ReactContext,
+} from './ReactTypes';
+
+export opaque type LanePriority =
+    | 0
+    | 1
+    | 2
+    | 3
+    | 4
+    | 5
+    | 6
+    | 7
+    | 8
+    | 9
+    | 10
+    | 11
+    | 12
+    | 13
+    | 14
+    | 15
+    | 16
+    | 17;
+export opaque type Lanes = number;
+export opaque type Lane = number;
+export opaque type LaneMap<T> = Array<T>;
+
+export type ReactPriorityLevel = 99 | 98 | 97 | 96 | 95 | 90;
+
+export type ContextDependency<T> = {
+  context: ReactContext<T>,
+  observedBits: number,
+  next: ContextDependency<mixed> | null,
+  ...
+};
+
+export type Dependencies = {
+  lanes: Lanes,
+  firstContext: ContextDependency<mixed> | null,
+  ...
+};
